+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Cron Tasks
 ### Changed
 - Creating and editing positions is now sensitive to access rights.
-- The logo model is now situated in the branding app
+- The logo model is now situated in the branding app.
+- Phone number has been added to the registration form.
 ### Fixed
 - Editing and creation rights in recruitment for officials
 - Allow submitting an application without changing a draft